requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
python-dotenv==1.0.0
newspaper3k==0.2.8
streamlit==1.31.1
nltk==3.8.1
google-generativeai==0.3.2
lxml==5.1.0
feedparser==6.0.10 
openai==1.68.2